CountDownTimerクラスを別ファイルに移行しました
