+++ HiitTimer/HiitTimer/TimerView.swift
@@ -7,7 +7,6 @@
 //
 
 import SwiftUI
-import Foundation
 
 struct TimerView: View {
     @ObservedObject var countdowntimer = CountDownTimer()
@@ -40,32 +39,6 @@ struct TimerView: View {
             // タイマーを表示
             Text("\(self.countdowntimer.counter)")
         }.font(.largeTitle)
-    }
-}
-
-// タイマークラスを定義
-class CountDownTimer: ObservableObject {
-    // 共有できる変数
-    @Published var counter: Int = 0
-    // タイマーを作成
-    var timer = Timer()
-    
-    // 実際に使うタイマーを作ってスタート
-    func start() {
-        timer = Timer.scheduledTimer(withTimeInterval: 1.0, repeats: true) {_ in
-            self.counter += 1
-        }
-    }
-    
-    // タイマーを止める
-    func stop() {
-        timer.invalidate()
-    }
-    
-    // タイマーをリセット
-    func reset() {
-        timer.invalidate()
-        counter = 0
     }
 }
